Use new code, delete old code
